+++ wdpostinit.sh
@@ -1,13 +1,15 @@
 #!/bin/bash
+
+#source support/support2.sh
 #
 # Main Control script for Free/TrueNAS CORE & SCALE on Western Digital PR2100?/PR4100 
 # Based off wdhws v1.0 by TFL (stefaang)
 #
-# wdpreinit V1.1 by Coltonton
-#    - Fixed Some Typos/Cleaned up while I was here
-#    - Added Support For TrueNAS Scale as well as TrueNAS CORE
-#    - More Comments = More Better
-# 
+# wdpreinit V1.2 beta by Coltonton @D3ADCRU7R
+#    - I'm back working on this!
+#    - Decided to skip finishing 1.1 and am just going to 1.2 to have more fun
+#    - Code being rebassed with functions moved to librarys
+#  
 # BSD 3 LICENSE (inherited from TFL)
 # Thanks unix stackexchange question 231975 & github user @stefaang
 
@@ -29,241 +31,274 @@
 #10-blue/off 11-blue/blue 12-blue/red 13-blue/purple 14-blue/green 15-blue/teal 16-blue/yellow 17-blue/White
 #18-purple/off 19-purple/blue 1A-purple/red 1B-purple/purple 1C-purple/green 1D-purple/teal 1E-purple/yellow 1F-purple/White
 
+
+source support/hwSupport.sh
+source support/pageSupport.sh
+###########################################################################
+#########################   DO NOT TOUCH   ################################
+###########################################################################
+declare -A hwSaftey=(   [fanSpeedMinimum]=35  # Minimum allowed fan speed expressed as percent
+                        [cpuOptimalTemp]=35   # Optimal (desired) temp for CPU (commie C degrees not freedom F :()
+                        [cpuMaxTemp]=80       # Maximum CPU temp before going full beans
+                        [diskMaxTemp]=40      # Maximum DISK temp before going full beans
+                        [pmcMaxTemp]=64       # Maximum PMC temp before going full beans
+                        [ramMaxTemp]=40       # Maximum RAM temp before going full beans
+                        [updateRate]=10       # How often in seconds to update hardware stats    
+                        [hwOverTempAlarm]=0  # Alarm for when a piece of hardware is over temp.
+)
+
 ###########################################################################
 #############################   VARS   ####################################
-###########################################################################
-minfanspeed=30    # Minimum fan speed in percent
-maxcputemp=80     # Maximum CPU temp before going full beans
-opptemp=35        # Optimal (desired) temp (commie C degrees not freedom F :( )
-tty=/dev/ttyS2    # Used to init variable, gets changed based on kernal in get_i2c_TTY()
+###########################################################################              
+declare -A hwOverTempArray=()   # Array containing all devices that are overtemp
+hwBayHD=()           # Array containing all the drives that are in a bay
+hwExtraHDD=()        # Array for extra hard drives
+
+declare -A sysInfo=([hwSystem]=Linux       # Used to init kernal variable, gets changed based on kernal in get_sys_info()
+                    [hwTTY]=/dev/cuau3     # Used to init tty variable, gets changed based on kernal in get_sys_info()
+                    [hwCPUCoreCount]=4     # Count of how many CPU cores there are
+)
+declare -A datetime=([Month]=01 
+                    [Day]=01   
+                    [Year]=00       
+                    [Hour]=12      
+                    [Minute]=00      
+                    [Second]=00
+)
 
 
 ###########################################################################
 #############################   FUNCS   ###################################
 ###########################################################################
-get_i2c_TTY(){
-    getVer=$(uname -s)           # Get Linux Kernal (Linux vs FreeBSD for TrueNas Scal/Core)
-    if [ $getVer == 'FreeBSD' ]  # If FreeBSD Free/TrueNAS Core
-    then
-        echo Found FreeBSD 
-        tty=/dev/cuau3             # FreeBSD uses /dev/cuau3 for i2C coms
-    elif [ $getVer == 'Linux' ]  # If Linux Free/TrueNAS Scale
-    then
-        echo Found Linux
-        tty=/dev/ttyS2             # Linux uses much cooler (telatype) /dev/ttyS2 for i2C coms
-    else                         # Just in case to catch wrong systems
-        echo ERROR: Detected Kernal Type Does Not Match Any Supported By This Program
-        echo Or there was an error
-        exit 
+
+check_for_dependencies(){   # Simple just-to-be-safe check that SMART Mon exists
+    depenflag=0
+    # S.M.A.R.T Drive Utilities
+    smartctl -v >/dev/null
+    if [[ $? != 1 ]]; then
+        printf "\n** SMART not installed please run - sudo apt install smartmontools ** \n\n "
+        (( depenflag += 1 ))
+    fi
+
+    # lm-Sensors (For temp sensor data )
+    sensors -v >/dev/null
+    if [[ $? != 0 ]]; then
+        printf "\n** lm-sensors not installed please run - sudo apt install smartmontools ** \n\n "
+        (( depenflag+=2 ))
+    fi
+    
+    if [ $depenflag -gt 0 ]; then
+        #printf "Would you like me to install? y/n:"
+        #read resp
+        #if [[ $resp == "y"] && [ $depenflag -gt 0 ]; then
+        #    if [ $depenflag -eq 1 ]; then
+        #        sudo apt install smartmontools
+        #    elif [ $depenflag -eq 2 ]; then
+        #        sudo apt install lm-sensors
+        #    elif [ $depenflag -eq 3 ]; then
+        #        sudo apt install smartmontools && sudo apt install lm-sensors
+        #    fi
+        #fi
+        printf "\n\n## PROGRAM TERMINATED ##\n\n"
+        exit
     fi
 }
 
-setup_tty() {
-    exec 4<$tty 5>$tty
+get_sys_info(){         # Get system info based off kernal, as BSD/LINUX has not the same commands
+    case "$( uname -s )" in        # Get Linux Kernal (Linux vs FreeBSD for TrueNas Scale/Core)
+        Linux*)  sysInfo[hwSystem]=Linux;;
+        *BSD)	 sysInfo[hwSystem]=BSD;;
+        Darwin*) sysInfo[hwSystem]=MacOS;;
+        CYGWIN*) sysInfo[hwSystem]=Cygwin;;
+        MINGW*)  sysInfo[hwSystem]=MinGw;;
+        *)       sysInfo[hwSystem]="Other"
+    esac
+    if [[ ! ${sysInfo[hwSystem]} =~ Linux|BSD ]]; then  # If system is not Linux or *BSD Show unsupported message
+        echo "Sorry, This software version for the WD PR4100 Hawrdware does not support $hwSystem platform."
+        echo "Please create an issue on Github to see about gettin support added"
+        exit 1
+    fi          
+    if [ "${sysInfo[hwSystem]}" == BSD ]; then      # If *BSD Free/TrueNAS Core
+        echo '# GETINFO: Detected BSD Kernal #'      # Show what kernal was identified
+        sysInfo[hwTTY]=/dev/cuau3                    # FreeBSD uses /dev/cuau3 for i2C comms to PR4100 front hardware
+        get_int_drives                      # Get location of ONLY internal bay drives
+        sysInfo[hwCPUCoreCount]=$(sysctl -n hw.ncpu) # Get how many CPU cores
+    elif [ ${sysInfo[hwSystem]} == Linux ]; then  # If Linux Free/TrueNAS Scale
+        echo '# GETINFO: Detected Linux Kernal #'    # Show what kernal was identified
+        sysInfo[hwTTY]=/dev/ttyS2                    # Linux uses much cooler (telatype) /dev/hwTTYS2 for i2C comms to PR4100 front hardware
+        get_int_drives                      # Get location of ONLY internal bay drives
+        sysInfo[hwCPUCoreCount]=$(nproc)             # Get how many CPU cores
+    fi
+    export sysInfo                          #Export sysInfo so our helper can pick it up
 }
 
-setup_i2c() {
-    # load kernel modules required for the temperature sensor on the RAM modules
-    kldload -n iicbus smbus smb ichsmb
-}
-
-send() {
-    setup_tty
-    # send a command to the PMC module and echo the answer
-    echo -ne "$1\r" >&5
-    read ans <&4
-    if [ "$ans" = "ALERT" ]; then
-        echo -ne ALERT >&2
-        exit 2 
-    else
-        # keep this for debugging failing commands
-        if [ "$ans" = "ERR" ] || [ -z "$ans" ]; then
-            echo "CMD $1 gives ERR at $2" >&2
-            send_empty
-            ans=$(send "$1" $(($2 + 1)))
-            #exit 1
+get_int_drives(){       # Basicly just gets the location of the internal bay HDD's
+    for file in /dev/disk/by-id/ata*       # With each HDD decice thats ata (Internal Sata)
+    do
+        if [[ $file != *"-part"* ]]; then  # Filter out '-part$' devices as they are the same
+            tmparr+=( $( ls -l "/dev/disk/by-id/ata-${file:20:100}" | awk '{print $11}' | cut -b 7-10 )  ) # Get the /dev location
+            readarray -t hwBayHD < <(for a in "${tmparr[@]}"; do echo "/dev/$a"; done | sort) # Sort
         fi
+    done
+    echo "# GETINFO: Detected internal bay drives: ${hwBayHD[@]} #"
+}
+
+setup_i2c() {           # load kernel modules required for the temperature sensor on the RAM modules, only avalible on FreeNAS CORE
+    if [ ${sysInfo[hwSystem]} == BSD ]; then
+        kldload -n iicbus smbus smb ichsmb
     fi
-    # only echo the result for retries ($2 not empty)
-    if [ ! -z $2 ]; then 
-        echo "CMD $1 gives '$ans' at $2" >&2 
-    fi 
-    echo $ans
-    send_empty
-    # deconstruct tty file pointers, otherwise this script breaks on sleep 
-    exec 4<&- 5>&-
 }
 
-send_empty() {
-    # send a empty command to clear the output
-    echo -ne "\r" >&5
-    read ignore <&4
+get_datetime() {        # Duh.... Sorry lol, easy time/date vars cuz cleaner
+    datetime[Month]=$(date +"%m")
+    datetime[Day]=$(date +"%d")
+    datetime[Year]=$(date +"%y")
+    datetime[Hour]=$(date +"%H")
+    datetime[Minute]=$(date +"%M")
+    datetime[Second]=$(date +"%S") 
 }
 
-get_ncpu() {
-    # get the number of CPUs
-    sysctl -n hw.ncpu
-}
-
-get_coretemp() {
-    # get the CPU temperature and strip of the Celsius
-    sysctl -n dev.cpu.$1.temperature | cut -d'.' -f1
-}
-
-get_disktemp() {
-    # get the disk $i temperature only if it is spinning
-    smartctl -n standby -A /dev/ada0 | grep Temperature_Celsius | awk '{print $10}'
-}
-
-get_ramtemp() {
-    # get the memory temperature from the I2C sensor
-    smbmsg -s 0x98 -c 0x0$1 -i 1 -F %d
-}
-
-get_pmc() {
-    # get a value from the PMC
-    # e.g. TMP returns TMP=25 --> 25
+get_pmc() {             # Requires input - Get a value from the PMC ex. inputing RPM gets fan0's rpm
     send $1 | cut -d'=' -f2
 }
 
-show_welcome() {
-    # set welcome message
-    # maximum  "xxx xxx xxx xxx " 
-    send   "LN1=    FreeNAS     "
-    send   "LN2=    Running     " 
+get_disktemp() {        # Requires input - Get the disks temperature only if it is active, else return status
+    drivesel=$1                       # For some reason I need this and cant put it in later? Makes i2c break somehow...
+    smartctl -n standby -A $drivesel > /dev/null # Run command to get disk status
+    getstatus=$(echo $?)              # Get drive exit status
+    if [ "$getstatus" == "0" ]; then  # If the status of the drive is active, get its temperature
+        smartctl -n standby -A $drivesel | grep Temperature_Celsius | awk '{print $10}'
+    else                              # If the status of the drive is not active, return the exit status of the drive. Maybe its asleep/standby                
+        return $getstatus
+    fi
 }
 
-led(){
-    #echo "PowerMode:$1 - PowerColor:$2 - UsbMode:$3 - UsbColor$4"
-    if [ "$1" == SOLID ]; then
-        send BLK=00
-        send PLS=00
-        if [ "$2" == BLU ]; then
-            send LED=01
-        elif [ "$2" == RED ]; then
-            send LED=02
-        elif [ "$2" == PUR ]; then
-            send LED=03
-        elif [ "$2" == GRE ]; then
-            send LED=04
-        elif [ "$2" == TEA ]; then
-            send LED=05
-        elif [ "$2" == YLW ]; then
-            send LED=06
-        elif [ "$2" == WHT ]; then
-            send LED=07
-        fi
-    fi
-    if [ "$1" == FLASH ]; then
-        send LED=00
-        send PLS=00
-        if [ "$2" == BLU ]; then
-            send BLK=01
-        elif [ "$2" == RED ]; then
-            send BLK=02
-        elif [ "$2" == PUR ]; then
-            send BLK=03
-        elif [ "$2" == GRE ]; then
-            send BLK=04
-        elif [ "$2" == TEA ]; then
-            send BLK=05
-        elif [ "$2" == YLW ]; then
-            send BLK=06
-        elif [ "$2" == WHT ]; then
-            send BLK=07
-        fi
-    fi
-
-    if [ "$1" == PULSE ]; then
-        send PLS=01
-        send LED=00
-        send BLK=00
+get_cpucoretemp() {     # Requires input - Get the specified CPU core temperature and strip off the celsius unit
+    if [ ${sysInfo[hwSystem]} == BSD ]; then       #FreeNAS Core
+        sysctl -n dev.cpu.$1.temperature | cut -d'.' -f1
+    elif [ ${sysInfo[hwSystem]} == Linux ]; then   # TrueNAS Scale
+        sensors | grep "Core $1" | awk '{print $3}' | cut -d'.' -f1 | cut -b 2-3
     fi  
 }
 
-show_ip() {
-    send "LN1=Interface re$1"
-    ip=$(ifconfig re$1 | grep inet | awk '{printf $2}')
-    send "LN2=$ip"
+get_ramtemp() {         # Requires input - Get the memory (ram) temperature from the I2C sensor FREENAS CORE ONLY
+    if [ ${sysInfo[hwSystem]} == BSD ]; then
+        smbmsg -s 0x98 -c 0x0$1 -i 1 -F %d
+    elif [ ${sysInfo[hwSystem]} == Linux ]; then
+        return "0"
+        # sensors | grep "temp1" | awk '{print $2}' | cut -d'.' -f1 | cut -b 2-3
+    fi
 }
 
-monitor() {
-    lvl="COOL"
-    cpumaxtmp=0
-    minfanspeed=30 #Percent
-    maxcputemp=80 
-    opptemp=35
-
+updateHW() {             # Main Function - Does all the hardware updating.
     # check RPM (fan may get stuck) and convert hex to dec
-    fan=$(get_pmc FAN)
-    rpm=$((0x$(get_pmc RPM)))
-    echo "Got rpm $rpm"
-    if [ "$rpm" != ERR ]; then
-        if [ "$rpm" -lt 400 ]; then
-            echo "WARNING: low RPM - $rpm - clean dust!"
-            led FLASH RED
+    #readfanpercent=$(get_pmc FAN)
+    rpmhex=$(get_pmc RPM)
+    rpmdec=$((0x$rpmhex))
+    #clear
+    echo "FAN 0 RPM: $rpmdec"
+    if [ "$rpmdec" != ERR ]; then
+        if [ "$rpmdec" -lt 400 ]; then
+            echo "FAN 0 RPM WARNING: low RPM - $rpmdec - clean dust!"
+            set_pwr_led FLASH RED
         fi
     fi
     
-    # check pmc  
-    tmp=$((0x$(get_pmc TMP)))
-    if [ "$tmp" -gt 64 ]; then
-        pmclvl="HOT"
+    # Check the Temperature of the PMC and convert to hex 
+    tmphex=$(get_pmc TMP)   # I dont need to do this in 2 steps but VSCode complains soooooo..... 2 steps wowww 
+    tmpdec=$((0x$tmphex))   # use tmpdec=$((0x$(get_pmc TMP))) if you want, it works, i just hate the 'error'
+    if [ "$tmpdec" -gt ${hwSaftey[pmcMaxTemp]} ]; then
+        echo "WARNING: PMC surpassed maximum (${hwSaftey[pmcMaxTemp]}°C), full throttle activated!"
+        hwSaftey[hwOverTempAlarm]=1 
+        #hwOverTempArray+=("PMC $tmp°C/$pmcMaxTemp°C")"
     fi
 
-    # check disks [adjust this for PR2100!!]
-    for i in 0 1 2 3 ; do
+    # Check the Hard Drive Temperature [adjust this for PR2100!!] (<- IDK what that means)
+    highestcpucoretemp=0
+    printf "|------ DISK TEMPS ------\n"
+    for i in "${hwBayHD[@]}" ; do
         tmp=$(get_disktemp $i)
-        echo "disk $i is $tmp"
-        if [ ! -z $tmp ] && [ "$tmp" -gt 40 ]; then
-            echo "Disk $i temperature is $tmp" 
-            lvl="HOT"
-        fi
+        waserror=$(echo $?)
+        if [ $waserror -ne "0" ]; then
+            if [ $waserror == 2 ]; then
+                ret=standby
+            else
+                ret=Error
+            fi
+            echo "| Drive ${i:5:5} is in $ret status"
+        else
+            echo "| Drive ${i:5:15} is $tmp °C"
+            if [ ! -z $tmp ] && [ $tmp -gt ${hwSaftey[diskMaxTemp]} ]; then
+                echo "| WARNING: CPU Core$i surpassed maximum (${hwSaftey[diskMaxTemp]}°C), full throttle activated!" 
+                hwSaftey[hwOverTempAlarm]=1
+                #hwOverTempArray+=("HDD$i $tmp°C/$hddMaxTemp°C")
+            fi
+        fi  
     done
+    printf "|------------------------\n"
     
-    #                                                         max-opperating=a   fullfan-minfan=b    b/a= fan percent per degree
-    # check cpu #max 80 #opp 35 1.5% for every degree above 30%      80-35=45         100-30=70     70/45=1.5   
-    for i in $(seq $(get_ncpu)); do
-        tmp=$(get_coretemp $((i-1)))
-        echo "cpu $i is $tmp"
-        if [ "$tmp" -gt 80 ]; then
-            echo "CPU $i temperature is $tmp"
-            lvl="HOT"
+
+    # Check the Temperature of the CPU
+    printf "|---- CPU CORE TEMPS ----\n"
+    for i in $(seq ${sysInfo[hwCPUCoreCount]}); do
+        tmp=$(get_cpucoretemp $((i-1)))
+        echo "| cpu core$i is $tmp °C"
+        if [ $tmp -gt ${hwSaftey[cpuMaxTemp]} ]; then
+            echo "| WARNING: CPU Core$i surpassed maximum (${hwSaftey[cpuMaxTemp]}°C), full throttle activated!"
+            #hwOverTempArray+=("CPU$i $tmp°C/$cpuMaxTemp°C")
+            hwSaftey[hwOverTempAlarm]=1
         fi
-        if [ $tmp -gt $cpumaxtmp ]; then
-            cpumaxtmp=$tmp
+        if [ $tmp -gt $highestcpucoretemp ]; then
+            highestcpucoretemp=$tmp
         fi
     done
-
-    echo "CPU max temp is $cpumaxtmp"
-    newtmp=$(($cpumaxtmp-$opptemp))
-    setspeed=$(($newtmp*2+$minfanspeed-5))
-    echo "speed should be: $setspeed%"
-    if [ $setspeed -lt $minfanspeed ]; then
-            setspeed=$minfanspeed
-            echo "Fan speed below minimum, bumping to $minfanspeed%..."
-    fi
+    printf "|------------------------\n"
+    #echo "Highest CPU core temp is $highestcpucoretemp °C"
+    #                                                       max-opperating=a   fullfan-minfan=b    b/a= fan percent per degree
+    #Max-80 Optimal-35 1.5% = for every degree above 30%      80-35=45         100-30=70             70/45=1.5   
+    newtmp=$(("$highestcpucoretemp"-"${hwSaftey[cpuOptimalTemp]}"))  #MaxTemp 
+    setspeed=$(("$newtmp"*2+"${hwSaftey[fanSpeedMinimum]}"-5))
  
-    # check ram
-    for i in 0 1; do
-        tmp=$(get_ramtemp $i)
-        echo "ram$i temp is $tmp"
-        if [ "$tmp" -gt 40 ]; then
-            echo "RAM$i temperature is $tmp"
-            lvl="HOT"
-        fi
-    done 
+    # Check the installed RAM Temperature
+    printf "|------ RAM TEMPS -------\n"
+    if [ ${sysInfo[hwSystem]} == BSD ]; then      # If *BSD Free/TrueNAS Core
+        for i in 0 1; do
+            tmp=$(get_ramtemp $i)
+            echo "| ram$i temp is $tmp °C"
+            if [ "$tmp" -gt ${hwSaftey[ramMaxTemp]} ]; then
+            #if [ "$tmp" -gt 0 ]; then
+                echo "| WARNING: RAM$i surpassed maximum (${hwSaftey[ramMaxTemp]}°C), full throttle activated!"
+                #hwOverTempArray+=("RAM $tmp°C/$ramMaxTemp°C")
+                hwSaftey[hwOverTempAlarm]=1
+            fi
+        done 
+    else 
+        echo "|"
+        echo "| Currently Unsupported"
+        echo "|"
+    fi
+    printf "|------------------------\n"
 
-    echo "Temperature LVL is $lvl"
-    if [ "$lvl" == HOT ] ; then
-        if [ "$fan" != 40 ]; then
-            send FAN=64
-            led FLASH RED
-        fi
+    if [ ${#hwOverTempArray[@]} -gt 0 ] || [ ${hwSaftey[hwOverTempAlarm]} -eq 1 ]; then
+        echo " WARNING: SYSTEM OVER LIMIT TEMPERATURE(s) FAN SET TO 100% "
+        echo "${#hwOverTempArray[@]}"
+        hwSaftey[hwOverTempAlarm]=1               # Flag System Over Temp-ed
+        #hwLastOverTemp=$(get_datetime) # Save the time when the system over temped
+        send FAN=64                     # Full Beans Fan 100%
+        set_pwr_led FLASH RED           # Flash Power LED RED to warn
+        #write_logdata                  # OOOOO am I leaking future stuff?! 
     else
-        send FAN=$setspeed 
+        if [ $setspeed -lt ${hwSaftey[fanSpeedMinimum]} ]; then
+            echo "Calculated fan speed below minimum allowed, bumping to ${hwSaftey[fanSpeedMinimum]}%..."
+            setspeed=${hwSaftey[fanSpeedMinimum]}  # Set the fan to the min allowed
+        else
+            echo "Setting fan speed to: $setspeed%"
+            send FAN=$setspeed         # Set fan to mathed speed if not overtemped
+        fi
     fi
 }
+
+
 
 check_btn_pressed() {
     btn=$(get_pmc ISR) 
@@ -296,19 +331,33 @@ check_btn_pressed() {
     esac        
 }
 
-init() {
-    get_i2c_TTY
-    setup_tty
-    setup_i2c
-
-    echo "Getting system status and firmware!"
-    send VER
-    send CFG 
-    send STA
-    led SOLID BLU
-    show_welcome
+preload(){
+    set_pwr_led FLASH YLW                     # Set the Power LED to flash yellow as visual indicator
+    setDisplay "wdhardware.sh" "Starting..."  # Set front panel LCD Line 1 and 2
 }
 
+init() {
+    get_sys_info
+    preload
+
+    #echo ${sysInfo[hwOverTempAlarm]}
+    ##sysInfo[hwSystem]="CHANGEDD"
+    #echo ${sysInfo[hwSystem]}
+
+    #echo $WDHardwareScriptOnline
+    #WDHardwareScriptOnline=1
+    #echo $WDHardwareScriptOnline
+
+    ##exit
+
+    check_for_dependencies
+    setup_i2c
+    StartWDHW
+    
+    setDisplay "TrueNAS" "Running"
+    set_pwr_led SOLID BLU
+    printf "# INIT DONE # \n\n"
+}
 
 ###########################################################################
 #############################   MAIN   ####################################
@@ -316,13 +365,17 @@ init() {
 init
 
 while true; do
-    # adjust fan speed every 30 seconds
-    monitor
+    # adjust fan speed every 10 seconds
+    #echo "SUB"
+    #echo $WDHardwareScriptOnline
+    #WDHardwareScriptOnline=1
+    updateHW
+    echo "Next temp update in ${hwSaftey[updateRate]} seconds"
+
 
     # check for button presses
-    for i in $(seq 10); do 
+    for i in $(seq ${hwSaftey[updateRate]}); do 
         sleep 1
         check_btn_pressed
     done
 done
-